+++ skrl/utils/model_instantiators/jax/__init__.py
@@ -3,6 +3,7 @@ from enum import Enum
 from skrl.utils.model_instantiators.jax.categorical import categorical_model
 from skrl.utils.model_instantiators.jax.deterministic import deterministic_model
 from skrl.utils.model_instantiators.jax.gaussian import gaussian_model
+from skrl.utils.model_instantiators.jax.multicategorical import multicategorical_model
 
 
 # keep for compatibility with versions prior to 1.3.0